refactor(NoDisplay): fix parameter type

+++ src/traffic_light/NoDisplay.h
@@ -4,5 +4,5 @@
 class NoDisplay : public IDisplay {
   public:
     void begin() override {}
-    void show(int) override {}
+    void show(uint8_t) override {}
 };
